requirements.txt
SpeechRecognition
pyaudio
requests
elevenlabs
openai
google-generativeai